requirements.txt
Django==1.9
djangorestframework==3.3.1
oauthlib==1.0.3
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.13
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
wheel==0.24.0
